temporary fix for Piccolo

+++ libfemto/drivers/ns16550a.c
@@ -29,7 +29,7 @@ enum {
     UART_LSR_EF   = 0x80,  /* Erroneous data in FIFO */
 };
 
-static volatile uint8_t *uart;
+static volatile uint32_t *uart;
 
 static void ns16550a_init()
 {
